Add files via upload
I.  Create a double array that contains at least 8 elements. The array should be initialized
    during declaration. 
II. Print the sum of the elements that are less then 10.0 in the array that you created. Use for-each loop 
    for the task. 
III. Write a print statement that prints the size of the array that you created using the array 
     reference variable.