Add files via upload
First, Creating an abstract class named Shape that contains two instance variables “dim1” and “dim2” 
of integer type and an abstract method named printArea(). Then Developing classes named Rectangle, 
Triangle and Circle which will be the subclasses of Shape class. Overriding the printArea() method 
and printing the area of the given shape.