Add files via upload
Create a superclass called "Person" and two subclasses, "Student" and 
"Teacher," that inherit from the "Person" class. Person class contains two 
protected instance variables name and age as well as a method which display
the name and age of a person. For the Student class there will be a private 
instance variable called studentId and a method which indicate that the student 
is studying. Again for the Teacher class add a private instance variable called 
subject and a method which indicate that the teacher is teaching