Add files via upload
Q: Create a class named BankAccount and a class named Main,
which will have your main() method. Then do the following tasks:
1. Update the BankAccount class :
i) Add 3 variables - Name(String), ID(int) and Balance(double).
ii) Create a constructor taking all 3 variables as parameters.
iii) Implement the toString() method to print the details of the
bank account object (show all 3 variables)
iv) Implement the Comparable interface and implement the
compareTo() method on the basis of ID
2. i) In the main() method of the Main class, create 5 BankAccount
objects with IDs 1,2,3,4,5 and with unique Name and Balance for
each account.
ii) Create an ArrayList of BankAccounts and add all the 5 objects
into the list.
iii) Use the Collections.sort() method to sort the ArrayList and
print the List
iv) Create a Comparator to compare the BankAccount objects on
the basis of Balance and use Collections.sort() method to sort
the ArrayList again and print the List.